refactoring preview carousel
